requirements.txt
opencv-python
ultralytics
google-generativeai
os